+++ src/Platforms/Iar/UtestPlatform.cpp
@@ -156,7 +156,7 @@ char* PlatformSpecificStrStr(const char* s1, const char* s2)
    return strstr((char*)s1, (char*)s2);
 }
 
-int PlatformSpecificVSNprintf(char *str, unsigned int size, const char* format, va_list args)
+int PlatformSpecificVSNprintf(char *str, size_t size, const char* format, va_list args)
 {
    return vsnprintf( str, size, format, args);
 }
